feat: validação lógica da reserva finalizada, provavelmente validações a 100% 🙏 (#46)

+++ trabalho-pratico/src/entities/passenger.c
@@ -6,14 +6,14 @@ struct passenger {
     char* document_number;
     char* first_name;
     char* last_name;
-    long dob;
+    int dob;
     char* nationality;
     char* gender;
     // email, phone, address, photo não são armazenados para já
 
 };
 
-Passenger* create_passenger(const char* document_number, const char* first_name, const char* last_name, long dob, const char* nationality, const char* gender) {
+Passenger* create_passenger(const char* document_number, const char* first_name, const char* last_name, int dob, const char* nationality, const char* gender) {
 
     Passenger* passenger = malloc(sizeof(Passenger));
     if (passenger == NULL) return NULL;
